+++ src/core/BUILD.gn
@@ -657,7 +657,6 @@ openthread_core_files = [
   "thread/network_data_leader.cpp",
   "thread/network_data_leader.hpp",
   "thread/network_data_leader_ftd.cpp",
-  "thread/network_data_leader_ftd.hpp",
   "thread/network_data_local.cpp",
   "thread/network_data_local.hpp",
   "thread/network_data_notifier.cpp",
